+++ CHANGELOG.md
@@ -25,6 +25,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Changed the expected format of data in the ``cray-product-catalog`` ConfigMap
+  so that multiple Docker images can be associated with a single product
+  version. Maintains backwards compatibility with previous ``cray-product-catalog``
+  data format.
+
 ## [1.1.0] - 2021-07-26
 
 ### Added